requirements.txt
faiss-cpu>=1.7.4
numpy>=1.23
scipy>=1.10
scikit-learn>=1.3
torch>=2.0
transformers>=4.41
tqdm>=4.66
pyyaml>=6.0
sentence-transformers>=3.0
streamlit>=1.32
plotly>=5.20
networkx>=3.0
pandas>=1.5
